Extracting out saved object manipulation
