+++ src/tabletop_simulator/importers/individual_card_importer.rb
@@ -1,23 +1,27 @@
 require "fileutils"
-require_relative "../card_format_translator.rb"
+require_relative "../modules.rb"
 
 module TabletopSimulator
   module Importers
     class IndividualCardImporter
       include CardFormatTranslator
+      include SavedObjectsManipulator
 
       BASE_PATH = %w(Cards).freeze
       INDIVIDUAL_CARDS_PATH = (BASE_PATH + %w(Individual)).freeze
 
-      def initialize(saved_objects_folder, base_directory)
-        @saved_objects_folder = saved_objects_folder
-        @base_directory = base_directory
+      def initialize(*saved_object_args)
+        initialize_directory_pointers(*saved_object_args)
       end
 
-      def import(stored_card)
-        card = TabletopSimulator::Models::Object.new(translate_card(stored_card))
+      def import(stored_cards)
+        clear_saved_objects(INDIVIDUAL_CARDS_PATH)
 
-        save_individual_card(card, stored_card.spec.title)
+        stored_cards.each do |stored_card|
+          card = TabletopSimulator::Models::Object.new(translate_card(stored_card))
+
+          save_individual_card(card, stored_card.spec.title)
+        end
       end
 
       private
@@ -27,15 +31,7 @@ module TabletopSimulator
       end
 
       def save_card(card, name, folders)
-        dir_path = path_in_saved_objects(folders)
-        FileUtils.mkdir_p(dir_path)
-
-        path = File.join(dir_path, "#{name}.json")
-        File.write(path, JSON.pretty_generate(card.to_h))
-      end
-
-      def path_in_saved_objects(folders)
-        File.join(@saved_objects_folder, @base_directory, folders)
+        save_file("#{name}.json", JSON.pretty_generate(card.to_h), folders)
       end
     end
   end
