+++ src/tabletop_simulator/modules.rb
@@ -1,0 +1,2 @@
+require_relative "modules/saved_object_manipulator.rb"
+require_relative "modules/card_format_translator.rb"
